mejoras componentes dinamicos

+++ CargaDBConFicheros/src/componentes/ComboBoxPersonalizado.java
@@ -4,32 +4,32 @@ import javax.swing.JComboBox;
 
 public class ComboBoxPersonalizado extends JComboBox<String> {
 
-	private int posicionX;
-	private int posicionY;
+	private double posicionX;
+	private double posicionY;
 
 	public ComboBoxPersonalizado() {
 		super();
 	}
 
-	public ComboBoxPersonalizado(int posicionX, int posicionY) {
+	public ComboBoxPersonalizado(double posicionX, double posicionY) {
 		super();
 		this.posicionX = posicionX;
 		this.posicionY = posicionY;
 	}
 
-	public int getPosicionX() {
+	public double getPosicionX() {
 		return posicionX;
 	}
 
-	public void setPosicionX(int posicionX) {
+	public void setPosicionX(double posicionX) {
 		this.posicionX = posicionX;
 	}
 
-	public int getPosicionY() {
+	public double getPosicionY() {
 		return posicionY;
 	}
 
-	public void setPosicionY(int posicionY) {
+	public void setPosicionY(double posicionY) {
 		this.posicionY = posicionY;
 	}
