+++ CargaDBConFicheros/src/componentes/TablaPersonalizado.java
@@ -4,32 +4,32 @@ import javax.swing.JTable;
 
 public class TablaPersonalizado extends JTable {
 
-	private int posicionX;
-	private int posicionY;
+	private double posicionX;
+	private double posicionY;
 
 	public TablaPersonalizado() {
 		super();
 	}
 
-	public TablaPersonalizado(int posicionX, int posicionY) {
+	public TablaPersonalizado(double posicionX, double posicionY) {
 		super();
 		this.posicionX = posicionX;
 		this.posicionY = posicionY;
 	}
 
-	public int getPosicionX() {
+	public double getPosicionX() {
 		return posicionX;
 	}
 
-	public void setPosicionX(int posicionX) {
+	public void setPosicionX(double posicionX) {
 		this.posicionX = posicionX;
 	}
 
-	public int getPosicionY() {
+	public double getPosicionY() {
 		return posicionY;
 	}
 
-	public void setPosicionY(int posicionY) {
+	public void setPosicionY(double posicionY) {
 		this.posicionY = posicionY;
 	}
